+++ CHANGELOG.md
@@ -11,6 +11,9 @@ All notable changes to this project will be documented in this file.
 - handle exceptions in afterEach after a test failure (fixes #110)
 - context silently ignored when using describe<Class> for subcontexts. (#104)
 
+### Changed
+
+- root-describe: replace boolean isolation parameter with nullable enum with just one entry "OFF"
 
 ## 0.7.2 "Rough Trade" - 2022-07-22